Pull out abstract button class
This allows us to change the code in one place to affect all buttons
that instaniate from NeuButton
